requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.12.1
Automat==22.10.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==4.0.0
charset-normalizer==2.0.8
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.9
django-autoslug==1.9.8
django-cors-headers==3.10.0
django-country-regions==0.0.1
django-environ==0.8.1
django-extensions==3.2.1
django-filter==21.1
django-multiselectfield==0.1.12
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
hurry.filesize==0.9
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.4
numpy==1.23.5
oauthlib==3.1.1
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.1.0
pyparsing==3.0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
redis==4.4.0
requests==2.26.0
requests-oauthlib==1.3.0

service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
toml==0.10.2
txaio==22.2.1
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.7
XlsxWriter==3.0.7
xmltodict==0.13.0
zope.interface==5.5.2
